services/backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
beanie==1.23.0
boto3==1.28.71
botocore==1.31.71
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.4
jmespath==1.0.1
lazy-model==0.2.0
motor==3.3.1
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
